move readme test to a separate file to avoid feature-interference between compiled binaries in target/debug
