+++ linera-service/tests/readme_test.rs
@@ -1,0 +1,102 @@
+// Copyright (c) Facebook, Inc. and its affiliates.
+// Copyright (c) Zefchain Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+#[cfg(feature = "aws")]
+use linera_views::test_utils::LocalStackTestContext;
+use once_cell::sync::Lazy;
+use std::io::Write;
+use tempfile::tempdir;
+use tokio::{process::Command, sync::Mutex};
+
+/// A static lock to prevent integration tests from running in parallel.
+static INTEGRATION_TEST_GUARD: Lazy<Mutex<()>> = Lazy::new(|| Mutex::new(()));
+
+#[test_log::test(tokio::test)]
+async fn test_examples_in_readme() -> std::io::Result<()> {
+    let _guard = INTEGRATION_TEST_GUARD.lock().await;
+
+    let dir = tempdir().unwrap();
+    let file = std::io::BufReader::new(std::fs::File::open("../README.md")?);
+    let mut quotes = get_bash_quotes(file)?;
+    // Check that we have the expected number of examples starting with "```bash".
+    assert_eq!(quotes.len(), 1);
+    let quote = quotes.pop().unwrap();
+
+    let mut test_script = std::fs::File::create(dir.path().join("test.sh"))?;
+    write!(&mut test_script, "{}", quote)?;
+
+    let status = Command::new("bash")
+        .current_dir("..") // root of the repo
+        .arg("-e")
+        .arg("-x")
+        .arg(dir.path().join("test.sh"))
+        .status()
+        .await?;
+    assert!(status.success());
+    Ok(())
+}
+
+#[allow(clippy::while_let_on_iterator)]
+fn get_bash_quotes(reader: impl std::io::BufRead) -> std::io::Result<Vec<String>> {
+    let mut result = Vec::new();
+    let mut lines = reader.lines();
+
+    while let Some(line) = lines.next() {
+        let line = line?;
+        if line.starts_with("```bash") {
+            let mut quote = String::new();
+            while let Some(line) = lines.next() {
+                let line = line?;
+                if line.starts_with("```") {
+                    break;
+                }
+                quote += &line;
+                quote += "\n";
+            }
+            result.push(quote);
+        }
+    }
+
+    Ok(result)
+}
+
+#[cfg(feature = "aws")]
+mod aws_test {
+    use super::*;
+
+    const ROCKSDB_STORAGE: &str = "--storage rocksdb:server_\"$I\"_\"$J\".db";
+    const DYNAMO_DB_STORAGE: &str = "--storage dynamodb:server-\"$I\":localstack";
+
+    const BUILD: &str = "cargo build";
+    const AWS_BUILD: &str = "cargo build --features aws";
+
+    #[test_log::test(tokio::test)]
+    async fn test_examples_in_readme_with_dynamo_db() -> anyhow::Result<()> {
+        let _guard = INTEGRATION_TEST_GUARD.lock().await;
+
+        let _localstack_guard = LocalStackTestContext::new().await?;
+        let dir = tempdir().unwrap();
+        let file = std::io::BufReader::new(std::fs::File::open("../README.md")?);
+        let mut quotes = get_bash_quotes(file)?;
+        // Check that we have the expected number of examples starting with "```bash".
+        assert_eq!(quotes.len(), 1);
+        let quote = quotes.pop().unwrap();
+        assert_eq!(quote.matches(ROCKSDB_STORAGE).count(), 1);
+        let quote = quote.replace(ROCKSDB_STORAGE, DYNAMO_DB_STORAGE);
+        let quote = quote.replace(BUILD, AWS_BUILD);
+
+        let mut test_script = std::fs::File::create(dir.path().join("test.sh"))?;
+        write!(&mut test_script, "{}", quote)?;
+
+        let status = Command::new("bash")
+            .current_dir("..") // root of the repo
+            .arg("-e")
+            .arg("-x")
+            .arg(dir.path().join("test.sh"))
+            .status()
+            .await?;
+        assert!(status.success());
+        Ok(())
+    }
+}
